Develop (#2)
* Fix #388: concave cd geometry errors

Address a special case where convex geometries result in a concave cd
geometry when combined. This showed up as geometry errors and could be
created using a radius-mismatched cylinder and sphere in a cd geometry.
Also fix a bug where a small step is taken in the cylinder or sphere
howfar() routine by boundaryTolerance/2 instead of boundaryTolerance,
causing checks against the boundaryTolerance to fail in the cd geometry.

* Fix #383: xyz geometry errors in cd and envelope

Resort to the boundaryTolerance parameter to avoid particles getting
stuck at edges when an xyz geometry serves as a cd base geometry, which
is subsequently inscribed in an envelope.

* Avoid infinite loop in conestack geometry

Resort to the boundaryTolerance parameter to avoid a rare but possible
infinite loop trap in simulations with a conestack geometry comprising
more than one layer, causing particles to get stuck on the surface
between layers in the conestack.

* Curb errors at box corners with strict inequality

Change comparison operators from inequalities to strict inequalities, in
order to avoid particles getting stuck at corners in an egs_box
geometry.

* Add decay summary output for radionuclide source

* Fix radionuclide source normalization

Set the statistical weight of a particle to zero when the energy is zero
during source generation in EGS_RadionuclideSource. Otherwise, extra
energy is deposited since the rest mass is added to the source particle
energy in the shower() routine.

Fix a bug where the branching ratio was not used for alpha records in
EGS_RadionuclideSource. Add a check to make sure the branching ratios
add up to 1 for the products of a given parent nuclide. If not, the
missing branch is assumed to be fission. Add support for P and Q shells
for the initial vacancies of relaxations from decays.

Change the relaxation emissions to use the intensity of each shell,
instead of the total intensity from the ENSDF file. This is necessary
since the shell intensities do not always add up to the total intensity.

Improve compatibility with NNDC (National Nuclear Data Center) ensdf
files, by allowing for the normalization to be neglected. Also improve
the robustness of checking comments for atomic relaxation data.

* Fix ambiguities and bugs in radionuclide source

Account for the uncertainties on the decay intensities when the total
decay intensity does not equal 100. Use the total intensity for internal
conversion, even if the individual shell intensities don't add up to
equal the total.

Improve the handling of internal transitions from unknown levels,
including fixing a bug where the total transition intensity was assigned
to the gamma intensity.

Add the option to model extra transitions when there is a mismatch
between the intensities in and out of a level in the daughter.

Fix a bug where internal pair intensity was read as P-shell internal
conversion intensity. Remove a duplicated decay in the Th-234 ensdf
file.

* Add radionuclide internal pair production sampling

Sample of internal pair production during internal transitions in the
radionuclide source model. This improves the accuracy of internal
transition decay rates. However, currently the electron-positron pair
following internal pair production is not produced when the event is
sampled. This is left for future work.

* Fix windows egs++ library compilation failure

Add missing EGS_EXPORT flags in egs_ensdf for the radionuclide source,
and fix a typo in the radiative splitting ausgab Makefile, which
prevented the egs++ library compiling on windows.

* Fix parallel data files combination on windows

Toggle between MergedChannels and SeparateChannels testing modes in
egs_configure. Parallel runs on Windows were failing to combine *.egsdat
files at the end of the simulation, because the system command exit
status check was failing, hence a dummy egs_system function always
returned success.

One can check for successful compilation only or also for successful
execution by checking the test code output against a reference value.
Depending what mode is needed the QProcess communication channels for
standard output and error can be merged or separated.

Modify the system test echo argument: use 'echo off' to test the system
command on Windows, instead of 'echo hallo' which interfered with the
test output, making the test fail when it actually succeeded.

* Fix #485: string size mismatch in mortran inputs

Increase character array size from 80 to 256 characters. This fixes a
bug where simulations would not read the density correction file input
if specified in a material file. This only occurred on certain compiler
versions, due to a string assignment mismatch. This bug only existed
since pull request #445.

* Remove egs_view compilation dependency on vector.h

Remove the include file vector.h from the egs_view ui file, since it is
not required and caused a compilation error. Compilation without

* Update egs_view references for the latest version

Update the getting started manual so that references to egs_view
correspond to the latest version. Similarly, update the egs_view
instructions in the egs++ doxygen documentation with new text and
figures.

* Reset mortran line length to 80 in get_inputs

Remove spurious space beyond the 80 character limit in
get_inputs.mortran, introduced by mistake in commit 0965496.

* Fix radionuclide auger emissions for EADL

Fix a bug where the auger electrons from atomic relaxations during
egs_radionuclide_source source generation were not produced. This did
not affect atomic relaxations in EGSnrc in general, only the
radionuclide source when it was set to EADL relaxation mode. The auger
electrons were excluded because the relaxation code expected ECUT to be
provided in kinetic energy instead of total energy.

* Improve NNDC BNL Nudat2 ensdf file compatibility

Add better support for Nudat2 ensdf files from the National Nuclear Data
Center (NNDC), Brookhaven National Laboratory (BNL). If the gamma
intensity is not provided, check for RI= tags or TI for transition
intensity. Relax line length requirements. Fix a bug where the internal
conversion coefficient was multiplied by the normalization.

* Add enhanced electromagnetic transport macros

This code implements charged particle transport in electric and magnetic
(EM) fields as described by Malkov and Rogers (Med. Phys. 43, 4447–4458,
2016 doi:10.1118/1.4954318). This code includes:

- Three-point numerical integration of the Lorentz force for PRESTA-II
- Two-point numerical integration for PRESTA-I
- Specialized magnetic field analytical transport in single scatter mode
- Boundary crossing adapted for EM fields
- Vacuum transport in EM fields

EEMF macros are provided as two files: one for general simulations, and
one specific to BEAMnrc simulations (required due to the way howfar and
hownear are defined in BEAMnrc).

Note that electric field transport has not been thoroughly tested and
verified at this point. Spatially varying fields are not explicitly
supported, however in principle small step sizes would admit spatial
variations.

A detailed report covering the implementation and use of these macros is
provided by the authors via the Carleton Laboratory for Radiotherapy
Physics: http://bit.ly/egsnrc-eemf

* Update the readme to link to the reddit page

* Revert 9f94ea4: avoid infinite loop in conestack

This reverts commit 9f94ea4. The original commit was intended to fix an
infinite loop in conetacks, but ended up introducing geometry errors in
long simulations of ionization chambers. Moreover, it appears that the
original issue was incidentally fixed by a different commit.